+++ lib/widgets/supervizorsko_mjesto_widget.dart
@@ -12,7 +12,7 @@ class SupervizorskoMjestoWidget extends StatelessWidget {
 
   final int id;
   final int index;
-  final int velicina;
+  final double velicina;
   final int ugao;
   //final Mjesto mjestoDat;
 
@@ -27,8 +27,7 @@ class SupervizorskoMjestoWidget extends StatelessWidget {
         Transform.rotate(
           angle: ugao * math.pi / 180,
           child: Icon(Icons.event_seat,
-              color: const Color.fromARGB(255, 88, 88, 88),
-              size: velicina.toDouble()),
+              color: const Color.fromARGB(255, 88, 88, 88), size: velicina),
         ),
         Positioned(
           right: 0,
